BCK-4 Creating repositories

+++ app/schemas/analitics.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, Field
 from datetime import date
 
@@ -8,7 +10,7 @@ class TransactionAnalysisItem(BaseModel):
     registered_users_count: int = Field(ge=0)
     registered_and_deposit_users_count: int = Field(ge=0)
     registered_and_not_rollbacked_deposit_users_count: int = Field(ge=0)
-    not_rollbacked_deposit_amount: float = Field(ge=0)
-    not_rollbacked_withdraw_amount: float = Field(ge=0)
+    not_rollbacked_deposit_amount: Decimal = Field(ge=0)
+    not_rollbacked_withdraw_amount: Decimal = Field(ge=0)
     transactions_count: int = Field(ge=0)
     not_rollbacked_transactions_count: int = Field(ge=0)
