+++ app/schemas/transaction.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from pydantic import BaseModel
 
@@ -7,13 +8,13 @@ from app.enums import CurrencyEnum, TransactionStatusEnum
 
 class RequestTransactionModel(BaseModel):
     currency: CurrencyEnum
-    amount: float
+    amount: Decimal
 
 
 class TransactionModel(BaseModel):
     id: int | None
     user_id: int | None = None
     currency: CurrencyEnum | None = None
-    amount: float | None = None
+    amount: Decimal | None = None
     status: TransactionStatusEnum | None = None
     created: datetime | None = None
